services/pdf-toolbox/requirements.txt
flask
flask-cors
pymupdf
pypdf
google-cloud-aiplatform
google-genai
python-dotenv
pillow
python-multipart
werkzeug
gunicorn
